HWRNG
This is a hardware random number generator, designed to be cheap, easy
to understand, and testable.

It is made from a 74HC14 hex inverter (the entropy source), a 74HC74
dual D-type flip-flop (one flip-flop is used as the entropy extractor),
and a cheap microcontroller such as the Atmel ATMega32U2.

The output bit-rate seems to be about 4,000 bits per second. It can be
driven faster, but this may result in autocorrelation.

The hex inverters are configured as two ring oscillators with three
inverters per ring. That makes the oscillator period roughly 120
nanoseconds, and it's frequency roughly 8 MHz. If the oscillator is
sampled at intervals of about 250 microseconds, then there will be about
2,000 cycles between samples, and this seems to be enough cycles to
accumulate enough phase jitter between samples that the samples are
uncorrelated.

There is a small amount of voodoo underlying the hardware design, which
I am not willing at present to explain.